services/user-service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-JWT-Extended
redis
python-dotenv
gunicorn